requirements.txt
aiogram==3.15.0
python-dotenv==0.21.0